Add files via upload
Every so often, I will post a number of Python exercises for you, so that if you are just learning Python or looking to gain enough mastery in Python, you can improve yourself. The exercises are at three levels: easy, medium and hard. and all of them are explained in the comments.
But if you have any questions, ask them and I will be happy to answer them.